still use uint8 for cpuload rate, enabled init test code

+++ src/types.go
@@ -27,8 +27,8 @@ type ProcLoad struct {
 }
 
 type CPUItem struct {
-	// normalize to 0-65535, stands for 0-100%
-	Rate_user, Rate_sys, Rate_iowait, Rate_idle uint16
+	// normalize to 0-255, stands for 0-100%
+	Rate_user, Rate_sys, Rate_iowait, Rate_idle uint8
 }
 
 type CPULoad struct {
